tests: the process-global script registrations split by subject
`tests/integration/elle_scripts.rs` was 2180 lines and 118 registrations, each a
comment saying what the script guards and a two-line body running it under a
process-global mode. The comments are the content, and at that length nobody
reads them to find the one that covers the shape they are working on.

The registrations now sit in seven modules named for what they guard —
dissolution, captures, frames, tailcalls, fibers, containers, and the modes that
answer to a backend toggle rather than to the guardfree oracle. Each is within
the reading budget. `elle_scripts.rs` keeps the prose about why these files
cannot ride the ordinary corpus runner, and the two helpers that spawn them.

Every registration is carried whole and none is added or dropped: 118 before,
118 after.

`prose.rs` picks up one repair on the way past. `header_doc_paths` trimmed
backticks and commas but not brackets, so a header citing a document
mid-sentence — `(docs/impl/x.md § …)` — read as a path beginning with an open
paren and reported the file as missing. Brackets are now trimmed with the rest
of the punctuation.
